Move svg wrapper to children
