+++ ui/src/features/annotator/annotations/annotations.component.tsx
@@ -1,6 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
+import { CSSProperties } from 'react';
+
 import { useAnnotator } from '../annotator-provider.component';
 import { useSelectedAnnotations } from '../select-annotation-provider.component';
 import { Annotation } from './annotation.component';
@@ -11,6 +13,17 @@ type AnnotationsProps = {
     height: number;
     isFocussed: boolean;
 };
+
+const DEFAULT_ANNOTATION_STYLES = {
+    fillOpacity: 0.4,
+    fill: 'var(--annotation-fill)',
+    stroke: 'var(--annotation-stroke)',
+    strokeLinecap: 'round',
+    strokeWidth: 'calc(1px / var(--zoom-scale))',
+    strokeDashoffset: 0,
+    strokeDasharray: 0,
+    strokeOpacity: 'var(--annotation-border-opacity, 1)',
+} satisfies CSSProperties;
 
 export const Annotations = ({ width, height, isFocussed }: AnnotationsProps) => {
     const { annotations } = useAnnotator();
@@ -23,10 +36,12 @@ export const Annotations = ({ width, height, isFocussed }: AnnotationsProps) => 
     ];
 
     return (
-        <MaskAnnotations annotations={orderedAnnotations} width={width} height={height} isEnabled={isFocussed}>
-            {orderedAnnotations.map((annotation) => (
-                <Annotation annotation={annotation} key={annotation.id} />
-            ))}
-        </MaskAnnotations>
+        <svg width={width} height={height} style={DEFAULT_ANNOTATION_STYLES}>
+            <MaskAnnotations annotations={orderedAnnotations} width={width} height={height} isEnabled={isFocussed}>
+                {orderedAnnotations.map((annotation) => (
+                    <Annotation annotation={annotation} key={annotation.id} />
+                ))}
+            </MaskAnnotations>
+        </svg>
     );
 };
